Add files via upload
Your Challenge:
You will only need to change the code in the HTML section of this Codepen, above the line with the comment that says "Above this is the section for you to edit".

For both the "col" class in the label element and the "col" class for the div that wraps around the select element, modify them so that they each take up 1/3rd of the row for small and up viewports.
Beneath the option "Blue", add two or more new color options of your choice. Make sure that the "value" holds a CSS color value such as "red" or "#fff".
Clicking on the label text "Color" should cause the Select input to become active, but currently, it does not. Can you figure out why, and how to fix it?
Verify that if you choose a color and click on the "Go" button, the background color changes to the color you selected.